KOKU-1355: Search in selaa tietopyynnot fix
KOKULT-344. The search in selaa tietopyynnöt doesn't work if the username
isn't found in ldap, an exception is raised which isn't handled. The surrounding
code already expects an empty string or a null in an error situation, so the fix
is just refactoring away the exception.

(Also trailing whitespace removal, use diff -b to ignore it when viewing the
changes in this commit)